[APM] Service profile UI prototype
Introduces a new route for querying profile data from Elasticsearch,
returning a set of aggregated call trees. Profile data are visualised
in a FlameGraph/icicle chart, using https://github.com/spiermar/d3-flame-graph.

Currently hard coded to use a terms aggregation with a maximum of
1000 unique stacks. We should probably use a composite aggregation,
otherwise we may miss nodes in the trees.

Currently hard coded to return CPU profile data only. This should be
controlled by the UI, and passed down as a query parameter.

Tooltips are a bit wonky, and should be replaced with something that
fits with the rest of the app.